requirements.txt
altgraph==0.17.4
certifi==2025.8.3
charset-normalizer==3.4.3
CTkListbox==1.5
CTkMessagebox==2.7
CTkToolTip==0.8
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging==25.0
pillow==11.3.0
pycairo==1.28.0
PyGObject==3.54.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
requests==2.32.5
setuptools==80.9.0
tkinterdnd2==0.4.3
urllib3==2.5.0
